+++ app/Http/Controllers/Api/ReportsPostsController.php
@@ -9,11 +9,12 @@ use App\Events\PostsReport;
 use Illuminate\Http\Request;
 use App\Traits\ResponseTrait;
 use App\Http\Controllers\Controller;
+use App\Traits\validator\ValidatorTrait;
 
 class ReportsPostsController extends Controller
 {
 
-     use ResponseTrait ;
+     use ResponseTrait , ValidatorTrait ;
 
      /** // ! only Admin can do that // 
      * todo  show all report posts .
@@ -30,15 +31,10 @@ class ReportsPostsController extends Controller
      */
     public function store(Request $request)
     {
-        //! rules
-        $rules = ["postid" => "required|numeric|exists:posts,id"];
         // ! valditaion
-        $validator = Validator::make($request->all(),$rules);
-
-        if($validator->fails()){
-            $code = $this->returnCodeAccordingToInput($validator);
-            return $this->returnValidationError($code,$validator);
-        }
+        $rules = ["postid" => "required|numeric|exists:posts,id"];
+        $validator = $this->validate($request,$rules);
+        if($validator !== true){return $validator;}
 
         $posts = Post::find($request->postid);
         $posts ->skip = false;
@@ -52,15 +48,10 @@ class ReportsPostsController extends Controller
      */
      public function skipreport(Request $request)
      {
-        //! rules
-        $rules = ["postid" => "required|numeric|exists:posts,id"];
         // ! valditaion
-        $validator = Validator::make($request->all(),$rules);
-    
-        if($validator->fails()){
-            $code = $this->returnCodeAccordingToInput($validator);
-            return $this->returnValidationError($code,$validator);
-        }
+        $rules = ["postid" => "required|numeric|exists:posts,id"];
+        $validator = $this->validate($request,$rules);
+        if($validator !== true){return $validator;}
 
         $posts = Post::find($request->postid);
         $posts ->skip = true;
@@ -74,6 +65,11 @@ class ReportsPostsController extends Controller
      */
     public function destroy(Request $request , $post)
     {
+        // ! valditaion
+        $rules = ["id" => "required|numeric|exists:posts,id"];
+        $validator = $this->validate($request,$rules);
+        if($validator !== true){return $validator;}
+
         // ? delete posts //
         $posts = Post::find($request->id) ;
         $posts->delete();
